Abstract `combine`, `aggregate_combined_blocks` into `TableBlockAccessor`, remove duplication
