+++ python/ray/data/_internal/arrow_block.py
@@ -449,82 +449,6 @@ class ArrowBlockAccessor(TableBlockAccessor):
 
         return find_partitions(table, boundaries, sort_key)
 
-    def combine(self, sort_key: "SortKey", aggs: Tuple["AggregateFn"]) -> Block:
-        """Combine rows with the same key into an accumulator.
-
-        This assumes the block is already sorted by key in ascending order.
-
-        Args:
-            sort_key: A column name or list of column names.
-            If this is ``None``, place all rows in a single group.
-
-            aggs: The aggregations to do.
-
-        Returns:
-            A sorted block of [k, v_1, ..., v_n] columns where k is the groupby
-            key and v_i is the partially combined accumulator for the ith given
-            aggregation.
-            If key is None then the k column is omitted.
-        """
-        keys: List[str] = sort_key.get_columns()
-
-        def iter_groups() -> Iterator[Tuple[Sequence[KeyType], Block]]:
-            """Creates an iterator over zero-copy group views."""
-            if not keys:
-                # Global aggregation consists of a single "group", so we short-circuit.
-                yield tuple(), self.to_block()
-                return
-
-            start = end = 0
-            iter = self.iter_rows(public_row_format=False)
-            next_row = None
-            while True:
-                try:
-                    if next_row is None:
-                        next_row = next(iter)
-                    next_keys = next_row[keys]
-                    while keys_equal(next_row[keys], next_keys):
-                        end += 1
-                        try:
-                            next_row = next(iter)
-                        except StopIteration:
-                            next_row = None
-                            break
-                    yield next_keys, self.slice(start, end)
-                    start = end
-                except StopIteration:
-                    break
-
-        builder = ArrowBlockBuilder()
-        for group_keys, group_view in iter_groups():
-            # Aggregate.
-            init_vals = group_keys
-            if len(group_keys) == 1:
-                init_vals = group_keys[0]
-
-            accumulators = [agg.init(init_vals) for agg in aggs]
-            for i in range(len(aggs)):
-                accumulators[i] = aggs[i].accumulate_block(accumulators[i], group_view)
-
-            # Build the row.
-            row = {}
-            if keys:
-                for k, gk in zip(keys, group_keys):
-                    row[k] = gk
-
-            count = collections.defaultdict(int)
-            for agg, accumulator in zip(aggs, accumulators):
-                name = agg.name
-                # Check for conflicts with existing aggregation name.
-                if count[name] > 0:
-                    name = self._munge_conflict(name, count[name])
-                count[name] += 1
-                row[name] = accumulator
-
-            builder.add(row)
-
-        return builder.build()
-
     @staticmethod
     def merge_sorted_blocks(
         blocks: List[Block], sort_key: "SortKey"
@@ -538,128 +462,6 @@ class ArrowBlockAccessor(TableBlockAccessor):
             blocks = TableBlockAccessor.normalize_block_types(blocks, "arrow")
             concat_and_sort = get_concat_and_sort_transform(DataContext.get_current())
             ret = concat_and_sort(blocks, sort_key)
-        return ret, ArrowBlockAccessor(ret).get_metadata(exec_stats=stats.build())
-
-    @staticmethod
-    def aggregate_combined_blocks(
-        blocks: List[Block],
-        sort_key: "SortKey",
-        aggs: Tuple["AggregateFn"],
-        finalize: bool,
-    ) -> Tuple[Block, BlockMetadata]:
-        """Aggregate sorted, partially combined blocks with the same key range.
-
-        This assumes blocks are already sorted by key in ascending order,
-        so we can do merge sort to get all the rows with the same key.
-
-        Args:
-            blocks: A list of partially combined and sorted blocks.
-            sort_key: The column name of key or None for global aggregation.
-            aggs: The aggregations to do.
-            finalize: Whether to finalize the aggregation. This is used as an
-                optimization for cases where we repeatedly combine partially
-                aggregated groups.
-
-        Returns:
-            A block of [k, v_1, ..., v_n] columns and its metadata where k is
-            the groupby key and v_i is the corresponding aggregation result for
-            the ith given aggregation.
-            If key is None then the k column is omitted.
-        """
-
-        stats = BlockExecStats.builder()
-        keys = sort_key.get_columns()
-
-        def _key_fn(r):
-            if keys:
-                return tuple(r[keys])
-            else:
-                return (0,)
-
-        # Replace `None`s and `np.nan` with NULL_SENTINEL to make sure
-        # we can order the elements (both of these are incomparable)
-        def safe_key_fn(r):
-            values = _key_fn(r)
-            return tuple(
-                [NULL_SENTINEL if v is None or is_nan(v) else v for v in values]
-            )
-
-        # Handle blocks of different types.
-        blocks = TableBlockAccessor.normalize_block_types(blocks, BlockType.ARROW)
-
-        iter = heapq.merge(
-            *[
-                ArrowBlockAccessor(block).iter_rows(public_row_format=False)
-                for block in blocks
-            ],
-            key=safe_key_fn,
-        )
-
-        next_row = None
-        builder = ArrowBlockBuilder()
-
-        while True:
-            try:
-                if next_row is None:
-                    next_row = next(iter)
-
-                next_keys = _key_fn(next_row)
-                next_key_columns = keys
-
-                def gen():
-                    nonlocal iter
-                    nonlocal next_row
-                    while keys_equal(_key_fn(next_row), next_keys):
-                        yield next_row
-                        try:
-                            next_row = next(iter)
-                        except StopIteration:
-                            next_row = None
-                            break
-
-                # Merge.
-                first = True
-                accumulators = [None] * len(aggs)
-                resolved_agg_names = [None] * len(aggs)
-                for r in gen():
-                    if first:
-                        count = collections.defaultdict(int)
-                        for i in range(len(aggs)):
-                            name = aggs[i].name
-                            # Check for conflicts with existing aggregation
-                            # name.
-                            if count[name] > 0:
-                                name = ArrowBlockAccessor._munge_conflict(
-                                    name, count[name]
-                                )
-                            count[name] += 1
-                            resolved_agg_names[i] = name
-                            accumulators[i] = r[name]
-                        first = False
-                    else:
-                        for i in range(len(aggs)):
-                            accumulators[i] = aggs[i].merge(
-                                accumulators[i], r[resolved_agg_names[i]]
-                            )
-                # Build the row.
-                row = {}
-                if keys:
-                    for col_name, next_key in zip(next_key_columns, next_keys):
-                        row[col_name] = next_key
-
-                for agg, agg_name, accumulator in zip(
-                    aggs, resolved_agg_names, accumulators
-                ):
-                    if finalize:
-                        row[agg_name] = agg.finalize(accumulator)
-                    else:
-                        row[agg_name] = accumulator
-
-                builder.add(row)
-            except StopIteration:
-                break
-
-        ret = builder.build()
         return ret, ArrowBlockAccessor(ret).get_metadata(exec_stats=stats.build())
 
     def block_type(self) -> BlockType:
